Modified relational diagram

+++ src/main/java/com/orar/Backend/Orar/service/SalaService.java
@@ -12,6 +12,8 @@ import com.orar.Backend.Orar.repository.SalaRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.List;
+
 @Service
 public class SalaService {
 
@@ -21,14 +23,21 @@ public class SalaService {
     @Autowired
     private CladireRepository cladireRepository;
 
+    public List<Sala> getAll() {
+        List<Sala> sali = salaRepository.findAll();
+        return sali;
+    }
+
     public Sala add(final SalaDTO salaDTO) throws SalaAlreadyExistsException, CladireNotFoundException {
         var newSala = buildSala(salaDTO);
         return add(newSala);
     }
+
     private Sala buildSala(final SalaDTO salaDTO) throws SalaAlreadyExistsException, CladireNotFoundException {
         checkSalaExists(salaDTO);
         return createSala(salaDTO);
     }
+
     private void checkSalaExists(final SalaDTO salaDTO) throws SalaAlreadyExistsException {
         salaRepository.findByNume(salaDTO.getNume())
                 .orElseThrow(() -> new SalaAlreadyExistsException("Sala already exists"));
@@ -42,6 +51,7 @@ public class SalaService {
     public Sala add(final Sala sala) {
         return salaRepository.save(sala);
     }
+
     private Sala createSala(final SalaDTO salaDTO) throws CladireNotFoundException {
         Sala sala = new Sala();
         sala.setNume(salaDTO.getNume());
@@ -53,15 +63,18 @@ public class SalaService {
         sala.setCladire(cladire);
         return sala;
     }
+
     public Sala update(final SalaDTO salaDTO) throws SalaNotFoundException, CladireNotFoundException {
         Sala existingSala = findSala(salaDTO);
         var updatedSala = updateSalaDetails(existingSala, salaDTO);
         return update(updatedSala);
     }
+
     private Sala findSala(final SalaDTO salaDTO) throws SalaNotFoundException {
         return salaRepository.findByNume(salaDTO.getNume())
                 .orElseThrow(() -> new SalaNotFoundException("Sala not found"));
     }
+
     private Sala updateSalaDetails(final Sala existingSala, final SalaDTO salaDTO) throws CladireNotFoundException {
         existingSala.setCapacitate(salaDTO.getCapacitate());
 
@@ -71,10 +84,15 @@ public class SalaService {
         existingSala.setCladire(cladire);
         return existingSala;
     }
+
     private Sala update(final Sala sala) {
         return salaRepository.save(sala);
     }
-    public void delete(final String numeSala) throws SalaNotFoundException {
+
+    public void delete(final String numeSala, final String numeCladire) throws SalaNotFoundException, CladireNotFoundException {
+        if (cladireRepository.findByNume(numeCladire).isEmpty()) {
+            throw new CladireNotFoundException("Cladire not found");
+        }
         Sala sala = salaRepository.findByNume(numeSala)
                 .orElseThrow(() -> new SalaNotFoundException("Sala not found"));
         salaRepository.delete(sala);
